Added empty reading ADC MCP3424

+++ Trik_Devices/trik_softi2c.h
@@ -35,7 +35,7 @@ struct tI2CRegisters
     uint8_t IERR;
     uint8_t ISTA;
     uint16_t IIDX;
-    uint16_t IVAL;
+    uint32_t IVAL;
 };
 
 //Registers
